requirements.txt
accelerate
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
antlr4-python3-runtime
asttokens
async-timeout
attrs
backcall
beautifulsoup4
biopython
brotlipy
certifi
cffi
charset-normalizer
click
colorama
comm
contourpy
cryptography
cycler
dataclasses
datasets
debugpy
decorator
dill
dmslogo
entrypoints
eval_type_backport
executing
fair-esm
filelock
fonttools
frozenlist
fsspec
gdown
gitdb
GitPython
gmpy2
graphviz
hf-xet
huggingface-hub
idna
iminuit
importlib-metadata
importlib_resources
ipykernel
ipython
ipywidgets
jedi
Jinja2
joblib
jupyter_client
jupyter_core
jupyterlab_widgets
kiwisolver
lightning-utilities
llvmlite
logomaker
lxml
MarkupSafe
matplotlib
matplotlib-inline
mkl-fft
mkl-random
mkl-service
mpmath
multidict
multiprocess
nest_asyncio
networkx
nltk
numba
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
omegaconf
OmegaFold
OpenAttack
packaging
palettable
pandas
parso
pexpect
pickleshare
Pillow
pip
platformdirs
polars
polyleven
prompt_toolkit
propcache
protobuf
psutil
ptyprocess
pure_eval
pyarrow
pycparser
pydantic
pydantic_core
Pygments
pynndescent
pyOpenSSL
pyparsing
PySocks
python-dateutil
pytorch-lightning
pytz
PyYAML
pyzmq
regex
requests
sacremoses
safetensors
scikit-learn
scipy
seaborn
sentry-sdk
setuptools
six
smmap
soupsieve
stack-data
sympy
threadpoolctl
tmscoring
tokenizers
torch
torchaudio
torchinfo
torchmetrics
torchvision
tornado
tqdm
traitlets
transformers
triton
typing_extensions
umap-learn
urllib3
wandb
wcwidth
wget
wheel
widgetsnbextension
xxhash
yarl
zipp
zstandard
